requirements.txt
streamlit>=1.10.0
<<<<<<< HEAD
PyGithub
=======
PyGithub>=1.55
controlflow>=0.1.0
google-cloud-aiplatform>=1.25.0
>>>>>>> 0f027eb8de8d5cabc450a62b4eb40906df016796
